moved hook to a functional component
